% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_views.R
\name{individuals}
\alias{individuals}
\title{Focus an idiostats fit on all individual models}
\usage{
individuals(x)
}
\arguments{
\item{x}{An idiostats fit.}
}
\value{
An idiostats fit view.
}
\description{
Focus an idiostats fit on all individual models
}
