% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_views.R
\name{overall}
\alias{overall}
\title{Focus an idiographic fit on overall metric rows}
\usage{
overall(x)
}
\arguments{
\item{x}{An idiographic fit.}
}
\value{
An idiographic fit view.
}
\description{
Focus an idiographic fit on overall metric rows
}
