% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_views.R
\name{people}
\alias{people}
\title{Focus an idiographic fit on selected people}
\usage{
people(x, ids)
}
\arguments{
\item{x}{An idiographic fit.}

\item{ids}{Person IDs.}
}
\value{
An idiographic fit view.
}
\description{
Focus an idiographic fit on selected people
}
