% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_views.R
\name{person}
\alias{person}
\title{Focus an idiographic fit on one person}
\usage{
person(x, id, ...)
}
\arguments{
\item{x}{An idiographic fit.}

\item{id}{Person ID.}

\item{...}{Ignored.}
}
\value{
An idiographic fit view.
}
\description{
Focus an idiographic fit on one person
}
