% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_views.R
\name{pooled}
\alias{pooled}
\title{Focus an idiographic fit on pooled models}
\usage{
pooled(x)
}
\arguments{
\item{x}{An idiographic fit.}
}
\value{
An idiographic fit view.
}
\description{
Focus an idiographic fit on pooled models
}
