% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{print.forecast_result}
\alias{print.forecast_result}
\title{Print method for forecast validation results}
\usage{
\method{print}{forecast_result}(x, ...)
}
\arguments{
\item{x}{A \code{forecast_result} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for forecast validation results
}
