% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso.R
\name{print.glasso_path_result}
\alias{print.glasso_path_result}
\title{Print a graphical-lasso path}
\usage{
\method{print}{glasso_path_result}(x, ...)
}
\arguments{
\item{x}{A \code{glasso_path_result} from \code{\link[=glasso_path]{glasso_path()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a graphical-lasso path
}
