% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_var_each.R
\name{print.gvar_list}
\alias{print.gvar_list}
\title{Print a list of per-subject graphical VARs}
\usage{
\method{print}{gvar_list}(x, ...)
}
\arguments{
\item{x}{A \code{gvar_list}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a list of per-subject graphical VARs
}
