% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_var.R
\name{print.gvar_result}
\alias{print.gvar_result}
\title{Print Method for gvar_result}
\usage{
\method{print}{gvar_result}(x, ...)
}
\arguments{
\item{x}{A \code{gvar_result} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object, invisibly.
}
\description{
Print Method for gvar_result
}
