% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{print.model_comparison}
\alias{print.model_comparison}
\title{Print method for model comparisons}
\usage{
\method{print}{model_comparison}(x, ...)
}
\arguments{
\item{x}{A \code{model_comparison} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for model comparisons
}
