% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlvar_bayes.R
\name{print.net_mlvar_bayes}
\alias{print.net_mlvar_bayes}
\title{Print method for net_mlvar_bayes}
\usage{
\method{print}{net_mlvar_bayes}(x, digits = 2, ...)
}
\arguments{
\item{x}{A \code{net_mlvar_bayes} object from \code{\link[=build_mlvar_bayes]{build_mlvar_bayes()}}.}

\item{digits}{Digits for printed network matrices.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for net_mlvar_bayes
}
