% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usem.R
\name{print.net_usem}
\alias{print.net_usem}
\title{Print method for uSEM fits}
\usage{
\method{print}{net_usem}(x, digits = 2, ...)
}
\arguments{
\item{x}{A \code{net_usem} object.}

\item{digits}{Number of digits used for printed network matrices.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for uSEM fits
}
