% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{print.preprocess_result}
\alias{print.preprocess_result}
\title{Print method for preprocessing results}
\usage{
\method{print}{preprocess_result}(x, ...)
}
\arguments{
\item{x}{A \code{preprocess_result} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for preprocessing results
}
