% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_graphical_var.R
\name{print.rolling_gvar_result}
\alias{print.rolling_gvar_result}
\title{Print method for rolling graphical VAR results}
\usage{
\method{print}{rolling_gvar_result}(x, ...)
}
\arguments{
\item{x}{A \code{rolling_gvar_result} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for rolling graphical VAR results
}
