% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_bayes.R
\name{print.var_bayes_result}
\alias{print.var_bayes_result}
\title{Print method for var_bayes_result}
\usage{
\method{print}{var_bayes_result}(x, digits = 2, ...)
}
\arguments{
\item{x}{A \code{var_bayes_result}.}

\item{digits}{Digits for printed networks.}

\item{...}{Unused.}
}
\value{
Invisibly \code{x}.
}
\description{
Print method for var_bayes_result
}
