% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_stats_identity.R
\name{summary.idiographic_fit}
\alias{summary.idiographic_fit}
\alias{summary.idiostats_fit}
\title{Summarise a consolidated fit}
\usage{
\method{summary}{idiographic_fit}(object, ...)

\method{summary}{idiostats_fit}(object, ...)
}
\arguments{
\item{object}{A consolidated \code{idiographic_fit}.}

\item{...}{Filters passed to \code{\link[=metrics]{metrics()}}.}
}
\value{
A model-metrics \code{data.frame}.
}
\description{
Summarise a consolidated fit
}
