% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_generics.R
\name{compare}
\alias{compare}
\title{Compare idiographic fits}
\usage{
compare(...)
}
\arguments{
\item{...}{One or more idiographic fits.}
}
\value{
A tidy data frame formed from each fit's overall metrics.
}
\description{
Compare idiographic fits
}
