% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_views.R
\name{individuals}
\alias{individuals}
\title{Focus an idiographic fit on all individual models}
\usage{
individuals(x)
}
\arguments{
\item{x}{An idiographic fit.}
}
\value{
An idiographic fit view.
}
\description{
Focus an idiographic fit on all individual models
}
