% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_views.R
\name{overall}
\alias{overall}
\title{Focus an idiostats fit on overall metric rows}
\usage{
overall(x)
}
\arguments{
\item{x}{An idiostats fit.}
}
\value{
An idiostats fit view.
}
\description{
Focus an idiostats fit on overall metric rows
}
