% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_views.R
\name{people}
\alias{people}
\title{Focus an idiostats fit on selected people}
\usage{
people(x, ids)
}
\arguments{
\item{x}{An idiostats fit.}

\item{ids}{Person IDs.}
}
\value{
An idiostats fit view.
}
\description{
Focus an idiostats fit on selected people
}
