% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_plots.R
\name{plot_importance}
\alias{plot_importance}
\title{Plot feature importance}
\usage{
plot_importance(x, model = NULL, scope = NULL, subject = NULL, n = 15, ...)
}
\arguments{
\item{x}{An idiographic fit.}

\item{model, scope, subject}{Optional filters.}

\item{n}{Number of rows/variables to plot.}

\item{...}{Passed to base plotting functions.}
}
\value{
Invisibly, the plotted importance table.
}
\description{
Plot feature importance
}
