% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_views.R
\name{pooled}
\alias{pooled}
\title{Focus an idiostats fit on pooled models}
\usage{
pooled(x)
}
\arguments{
\item{x}{An idiostats fit.}
}
\value{
An idiostats fit view.
}
\description{
Focus an idiostats fit on pooled models
}
