% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_var.R
\name{print.rolling_var_result}
\alias{print.rolling_var_result}
\title{Print method for rolling VAR results}
\usage{
\method{print}{rolling_var_result}(x, ...)
}
\arguments{
\item{x}{A \code{rolling_var_result} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for rolling VAR results
}
