% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{print.stability_result}
\alias{print.stability_result}
\title{Print method for stability results}
\usage{
\method{print}{stability_result}(x, ...)
}
\arguments{
\item{x}{A \code{stability_result} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for stability results
}
