% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{print.var_list}
\alias{print.var_list}
\title{Print a list of per-subject ordinary VARs}
\usage{
\method{print}{var_list}(x, ...)
}
\arguments{
\item{x}{A \code{var_list}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a list of per-subject ordinary VARs
}
